+++ fla/ops/triton/triton_core/kda/_kda_common/chunk_h.py
@@ -12,7 +12,7 @@ import triton.language as tl
 
 from fla.ops.triton.triton_core.index import prepare_chunk_offsets
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp
-from fla.ops.triton.triton_core.kda._kda_utils.utils import autotune_cache_kwargs, check_shared_mem
+from fla.ops.triton.triton_core.utils import autotune_cache_kwargs, check_shared_mem
 
 BKV_LIST = [32, 64] if check_shared_mem() else [16, 32]
