--- fla/ops/triton/triton_core/kda/_kda_utils/constant.py
@@ -1,11 +1,0 @@
-# Copyright © 2026 Huawei Technologies Co., Ltd.
-# Based on flash-linear-attention: https://github.com/fla-org/flash-linear-attention
-#
-# This file contains code copied and/or modified from the flash-linear-attention project.
-# The original source code was licensed under the MIT license and included
-# the following copyright notice:
-# Copyright (c) 2023-2026, Songlin Yang, Yu Zhang, Zhiyuan Li
-
-# Approximate value of 1/ln(2), used for log/exp base conversion
-# Best FP32 approximation: 1.4426950216 (hex 0x3FB8AA3B)
-RCP_LN2 = 1.4426950216
