+++ fla/ops/triton/triton_core/kda/_kda_utils/l2norm.py
@@ -11,7 +11,7 @@ import torch.nn as nn
 import triton
 import triton.language as tl
 
-from fla.ops.triton.triton_core.kda._kda_utils.utils import autotune_cache_kwargs, input_guard
+from fla.ops.triton.triton_core.utils import autotune_cache_kwargs, input_guard
 
 BT_LIST = [8, 16, 32, 64, 128]
 NUM_WARPS_AUTOTUNE = [1, 2, 4, 8, 16, 32]
