+++ fla/ops/triton/triton_core/kda/chunk_bwd.py
@@ -23,7 +23,7 @@ from fla.ops.triton.triton_core.cumsum import chunk_local_cumsum
 from fla.ops.triton.triton_core.index import prepare_chunk_indices
 from fla.ops.triton.triton_core.constant import RCP_LN2
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
-from fla.ops.triton.triton_core.kda._kda_utils.utils import (
+from fla.ops.triton.triton_core.utils import (
     autotune_cache_kwargs,
 )
