+++ fla/ops/triton/triton_core/kda/chunk_intra_token_parallel.py
@@ -11,7 +11,7 @@ import triton
 import triton.language as tl
 
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
-from fla.ops.triton.triton_core.kda._kda_utils.utils import autotune_cache_kwargs
+from fla.ops.triton.triton_core.utils import autotune_cache_kwargs
 
 @triton.heuristics({
     'IS_VARLEN': lambda args: args['cu_seqlens'] is not None,
