+++ fla/ops/triton/triton_core/kda/fused_recurrent.py
@@ -12,7 +12,7 @@ import triton.language as tl
 
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp
 from fla.ops.triton.triton_core.softplus import softplus
-from fla.ops.triton.triton_core.kda._kda_utils.utils import input_guard
+from fla.ops.triton.triton_core.utils import input_guard
 
 @triton.heuristics(
     {
