+++ fla/ops/triton/triton_core/kda/wy_fast.py
@@ -12,7 +12,7 @@ import triton.language as tl
 
 from fla.ops.triton.triton_core.index import prepare_chunk_indices
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
-from fla.ops.triton.triton_core.kda._kda_utils.utils import autotune_cache_kwargs, check_shared_mem
+from fla.ops.triton.triton_core.utils import autotune_cache_kwargs, check_shared_mem
 
 @triton.heuristics({
     'STORE_QG': lambda args: args['qg'] is not None,
